Docstrings, moving chsh into two_qubit_interactions and update example runcard
