+++ src/qibocal/protocols/characterization/__init__.py
@@ -3,7 +3,6 @@ from enum import Enum
 from .allxy.allxy import allxy
 from .allxy.allxy_drag_pulse_tuning import allxy_drag_pulse_tuning
 from .allxy.drag_pulse_tuning import drag_pulse_tuning
-from .chsh import chsh_circuits, chsh_pulses
 from .classification import single_shot_classification
 from .coherence.spin_echo import spin_echo
 from .coherence.t1 import t1
@@ -31,8 +30,7 @@ from .resonator_punchout_attenuation import resonator_punchout_attenuation
 from .resonator_spectroscopy import resonator_spectroscopy
 from .resonator_spectroscopy_attenuation import resonator_spectroscopy_attenuation
 from .signal_experiments.time_of_flight_readout import time_of_flight_readout
-from .two_qubit_interaction.chevron import chevron
-from .two_qubit_interaction.cz_virtualz import cz_virtualz
+from .two_qubit_interaction import chevron, chsh_circuits, chsh_pulses, cz_virtualz
 
 
 class Operation(Enum):
