+++ src/qibocal/protocols/characterization/two_qubit_interaction/chsh/__init__.py
