+++ src/qibocal/protocols/characterization/two_qubit_interaction/chsh/circuits.py
