+++ src/qibocal/protocols/characterization/two_qubit_interaction/chsh/pulses.py
